♻️ refactor: split session tests into multiple files
