--- tests/test_session.py
@@ -1,2138 +1,0 @@
-"""Tests for SessionManager (no LLM tokens needed)."""
-
-from __future__ import annotations
-
-import asyncio
-from datetime import UTC, datetime, timedelta
-from decimal import Decimal
-from pathlib import Path
-from typing import Any
-from unittest.mock import AsyncMock, MagicMock, patch
-
-import pytest
-from pydantic_ai import ApprovalRequired
-from pydantic_ai.exceptions import UsageLimitExceeded
-from pydantic_ai.messages import ModelRequest, ModelResponse, TextPart, ToolCallPart, ToolReturnPart, UserPromptPart
-from pydantic_ai.models.test import TestModel
-from pydantic_ai.usage import RunUsage
-
-import carapace.security as security_mod
-import carapace.usage as usage_mod
-from carapace.bootstrap import ensure_data_dir
-from carapace.config import load_config
-from carapace.credentials import CredentialRegistry
-from carapace.git.store import GitStore
-from carapace.models import (
-    ContextGrant,
-    CredentialRegistryProtocol,
-    SessionAttributes,
-    SessionBudget,
-    SkillCredentialDecl,
-    ToolResult,
-)
-from carapace.sandbox.manager import SandboxManager
-from carapace.sandbox.state import SessionSandboxSnapshot
-from carapace.security.context import (
-    SentinelVerdict,
-    SessionSecurity,
-    ToolCallEntry,
-    UserEscalationDecision,
-    format_denial_message,
-    normalize_optional_message,
-)
-from carapace.security.sentinel import Sentinel
-from carapace.session import SessionEngine, SessionManager
-from carapace.session.turns import _non_slash_user_message_count
-from carapace.skills import SkillRegistry
-from carapace.usage import LlmRequestRecord, LlmRequestState, ModelUsage, SessionBudgetExceededError
-from carapace.ws_models import ApprovalRequest, TurnUsage
-
-
-def _patch_sentinel():
-    """Patch Sentinel class so its instances pass isinstance checks."""
-    mock_cls = MagicMock()
-    mock_cls.return_value = MagicMock(spec=Sentinel)
-    return patch("carapace.session.engine.Sentinel", mock_cls)
-
-
-def _without_timestamp(event: dict[str, Any]) -> dict[str, Any]:
-    return {key: value for key, value in event.items() if key != "timestamp"}
-
-
-def _without_timestamps(events: list[dict[str, Any]]) -> list[dict[str, Any]]:
-    return [_without_timestamp(event) for event in events]
-
-
-def test_create_session(tmp_path: Path):
-    mgr = SessionManager(tmp_path)
-    state = mgr.create_session()
-    assert len(state.session_id) == 25  # 2026-03-08-10-22-abcd1234
-    assert state.channel_type == "cli"
-    assert state.attributes == SessionAttributes()
-
-
-def test_create_session_persists_budget(tmp_path: Path):
-    mgr = SessionManager(tmp_path)
-    budget = SessionBudget(input_tokens=1_000, cost_usd=Decimal("5.00"), tool_calls=4)
-    state = mgr.create_session(budget=budget)
-
-    resumed = mgr.resume_session(state.session_id)
-    assert resumed is not None
-    assert resumed.budget.input_tokens == 1_000
-    assert resumed.budget.cost_usd == Decimal("5.00")
-    assert resumed.budget.tool_calls == 4
-
-
-def test_create_session_persists_private_attribute(tmp_path: Path):
-    mgr = SessionManager(tmp_path)
-    state = mgr.create_session(private=True)
-
-    resumed = mgr.resume_session(state.session_id)
-    assert resumed is not None
-    assert resumed.attributes.private is True
-
-
-def test_resume_session(tmp_path: Path):
-    mgr = SessionManager(tmp_path)
-    state = mgr.create_session()
-    resumed = mgr.resume_session(state.session_id)
-    assert resumed is not None
-    assert resumed.session_id == state.session_id
-
-
-def test_update_events_timestamps_inserted_and_replaced_entries(tmp_path: Path):
-    mgr = SessionManager(tmp_path)
-    state = mgr.create_session()
-    mgr.save_events(
-        state.session_id,
-        [
-            {"role": "user", "content": "first", "timestamp": "2026-01-01T00:00:00+00:00"},
-            {"role": "assistant", "content": "second", "timestamp": "2026-01-01T00:00:01+00:00"},
-        ],
-    )
-
-    def _mutate(events: list[dict[str, Any]]) -> None:
-        events.insert(0, {"role": "thinking", "content": "inserted"})
-        events[2] = {"role": "assistant", "content": "replaced"}
-
-    mgr.update_events(state.session_id, _mutate)
-
-    events = mgr.load_events(state.session_id)
-    assert events[0]["role"] == "thinking"
-    assert "timestamp" in events[0]
-    assert events[1]["timestamp"] == "2026-01-01T00:00:00+00:00"
-    assert events[2]["role"] == "assistant"
-    assert events[2]["content"] == "replaced"
-    assert "timestamp" in events[2]
-
-
-def test_resume_nonexistent(tmp_path: Path):
-    mgr = SessionManager(tmp_path)
-    assert mgr.resume_session("doesnotexist") is None
-
-
-def test_list_sessions(tmp_path: Path):
-    mgr = SessionManager(tmp_path)
-    s1 = mgr.create_session()
-    s2 = mgr.create_session()
-    sessions = mgr.list_sessions()
-    assert s1.session_id in sessions
-    assert s2.session_id in sessions
-
-
-def test_save_and_resume_state(tmp_path: Path):
-    from carapace.models import ContextGrant, SkillCredentialDecl
-
-    mgr = SessionManager(tmp_path)
-    state = mgr.create_session()
-    state.context_grants["my-skill"] = ContextGrant(
-        skill_name="my-skill",
-        domains={"example.com"},
-        credential_decls=[SkillCredentialDecl(vault_path="dev/test", description="test cred")],
-    )
-    mgr.save_state(state)
-
-    resumed = mgr.resume_session(state.session_id)
-    assert resumed is not None
-    assert "my-skill" in resumed.context_grants
-    assert "dev/test" in resumed.context_grants["my-skill"].vault_paths
-
-
-def test_save_and_load_llm_request_state(tmp_path: Path) -> None:
-    mgr = SessionManager(tmp_path)
-    state = mgr.create_session()
-    activity = LlmRequestState(
-        request_id="req-1",
-        source="agent",
-        model_name="anthropic:claude-haiku-4-5",
-        started_at=datetime.now(tz=UTC),
-        phase="processing_prompt",
-    )
-
-    mgr.save_llm_request_state(state.session_id, activity)
-
-    reloaded = mgr.load_llm_request_state(state.session_id)
-    assert reloaded is not None
-    assert reloaded.request_id == "req-1"
-    assert reloaded.phase == "processing_prompt"
-
-    mgr.clear_llm_request_state(state.session_id)
-    assert mgr.load_llm_request_state(state.session_id) is None
-
-
-def test_save_and_load_sandbox_snapshot(tmp_path: Path) -> None:
-    mgr = SessionManager(tmp_path)
-    state = mgr.create_session()
-    snapshot = SessionSandboxSnapshot(
-        exists=True,
-        runtime="kubernetes",
-        status="scaled_down",
-        resource_id="carapace-sandbox-abc-0",
-        resource_kind="statefulset",
-        storage_present=True,
-        provisioned_bytes=1_073_741_824,
-        last_measured_used_bytes=123_456,
-        last_measured_at=datetime.now(tz=UTC),
-        updated_at=datetime.now(tz=UTC),
-    )
-
-    mgr.save_sandbox_snapshot(state.session_id, snapshot)
-
-    reloaded = mgr.load_sandbox_snapshot(state.session_id)
-    assert reloaded is not None
-    assert reloaded.runtime == "kubernetes"
-    assert reloaded.status == "scaled_down"
-    assert reloaded.last_measured_used_bytes == 123_456
-
-    mgr.clear_sandbox_snapshot(state.session_id)
-    assert mgr.load_sandbox_snapshot(state.session_id) is None
-
-
-def test_normalize_optional_message_strips_blank_values() -> None:
-    assert normalize_optional_message(None) is None
-    assert normalize_optional_message("   ") is None
-    assert normalize_optional_message("  blocked by user  ") == "blocked by user"
-
-
-def test_format_denial_message_includes_source_and_message() -> None:
-    assert format_denial_message("sentinel", "dangerous command") == (
-        "Sentinel denied this operation. dangerous command"
-    )
-    assert format_denial_message("user", None) == "User denied this operation."
-
-
-def test_missing_user_escalation_callback_denies() -> None:
-    from carapace.security.context import SessionSecurity
-
-    security = SessionSecurity("session-1")
-    result = asyncio.run(security.escalate_to_user("example.com", {"kind": "domain_access"}))
-    assert result == UserEscalationDecision(allowed=False)
-
-
-@pytest.mark.anyio
-async def test_skill_activation_inputs_use_context_grant(tmp_path: Path):
-    """Automatic skill setup reuses approved env/file credentials from the persisted context grant."""
-    skill_name = "reinject-skill"
-    skill_dir = tmp_path / "skills" / skill_name
-    skill_dir.mkdir(parents=True)
-    (skill_dir / "SKILL.md").write_text(f"---\nname: {skill_name}\n---\n")
-    (skill_dir / "carapace.yaml").write_text(
-        "credentials:\n"
-        "  - vault_path: vault/secret\n"
-        "    description: API key\n"
-        "    env_var: API_KEY\n"
-        "    file: .secrets/key.txt\n"
-    )
-
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    state = engine.session_mgr.create_session()
-    sid = state.session_id
-    active = engine.get_or_activate(sid)
-    active.state.context_grants[skill_name] = ContextGrant(
-        skill_name=skill_name,
-        credential_decls=[
-            SkillCredentialDecl(
-                vault_path="vault/secret",
-                description="API key",
-                env_var="API_KEY",
-                file=".secrets/key.txt",
-            ),
-        ],
-    )
-
-    mock_reg = AsyncMock(spec=CredentialRegistryProtocol)
-    mock_reg.fetch = AsyncMock(return_value="secret-value")
-    engine._credential_registry = mock_reg
-
-    result = await engine._skill_activation_inputs(sid, skill_name)
-    assert result.environment == {"API_KEY": "secret-value"}
-    assert [(cred.path, cred.value) for cred in result.file_credentials] == [(".secrets/key.txt", "secret-value")]
-    mock_reg.fetch.assert_awaited_once_with("vault/secret")
-
-    active.state.context_grants.pop(skill_name, None)
-    mock_reg.fetch.reset_mock()
-    empty = await engine._skill_activation_inputs(sid, skill_name)
-    assert empty.environment == {}
-    assert empty.file_credentials == []
-    mock_reg.fetch.assert_not_called()
-
-    # Reload from disk when session is not active (same path as idle resume + venv sync)
-    active.state.context_grants[skill_name] = ContextGrant(
-        skill_name=skill_name,
-        credential_decls=[
-            SkillCredentialDecl(
-                vault_path="vault/secret",
-                description="API key",
-                env_var="API_KEY",
-                file=".secrets/key.txt",
-            ),
-        ],
-    )
-    engine.session_mgr.save_state(active.state)
-    engine.deactivate(sid)
-    mock_reg.fetch.reset_mock()
-    mock_reg.fetch = AsyncMock(return_value="from-disk")
-    result = await engine._skill_activation_inputs(sid, skill_name)
-    assert result.environment == {"API_KEY": "from-disk"}
-    assert [(cred.path, cred.value) for cred in result.file_credentials] == [(".secrets/key.txt", "from-disk")]
-
-
-# ---------------------------------------------------------------------------
-# SessionEngine: on_user_message from_self tests
-# ---------------------------------------------------------------------------
-
-
-class _FakeSubscriber:
-    """Minimal subscriber that records calls."""
-
-    def __init__(self) -> None:
-        self.user_messages: list[tuple[str, bool]] = []
-        self.token_chunks: list[str] = []
-        self.thinking_chunks: list[str] = []
-        self.errors: list[str] = []
-        self.error_events: list[tuple[str, bool]] = []
-        self.cancelled: int = 0
-        self.done_messages: list[tuple[str, TurnUsage]] = []
-        self.title_updates: list[tuple[str, TurnUsage | None]] = []
-        self.llm_activity_updates: list[LlmRequestState | None] = []
-
-    async def on_user_message(self, content: str, *, from_self: bool) -> None:
-        self.user_messages.append((content, from_self))
-
-    async def on_tool_call(self, tool: str, args: dict[str, Any], detail: str) -> None:
-        pass
-
-    async def on_tool_result(self, result: ToolResult) -> None:
-        pass
-
-    async def on_token(self, content: str) -> None:
-        self.token_chunks.append(content)
-
-    async def on_thinking_token(self, content: str) -> None:
-        self.thinking_chunks.append(content)
-
-    async def on_done(self, content: str, usage: TurnUsage) -> None:
-        self.done_messages.append((content, usage))
-
-    async def on_error(self, detail: str, *, turn_terminal: bool = False) -> None:
-        self.errors.append(detail)
-        self.error_events.append((detail, turn_terminal))
-
-    async def on_cancelled(self) -> None:
-        self.cancelled += 1
-
-    async def on_approval_request(self, req: ApprovalRequest) -> None:
-        pass
-
-    async def on_proxy_approval_request(self, request_id: str, domain: str, command: str) -> None:
-        pass
-
-    async def on_title_update(self, title: str, usage: TurnUsage | None = None) -> None:
-        self.title_updates.append((title, usage))
-
-    async def on_llm_activity(self, activity: LlmRequestState | None) -> None:
-        self.llm_activity_updates.append(activity)
-
-    async def on_domain_info(self, domain: str, detail: str) -> None:
-        pass
-
-
-def _make_engine(tmp_path: Path) -> SessionEngine:
-    ensure_data_dir(tmp_path)
-    config = load_config(tmp_path)
-    session_mgr = SessionManager(tmp_path)
-    registry = SkillRegistry(tmp_path / "skills")
-    skill_catalog = registry.scan()
-    sandbox_mgr = MagicMock(spec=SandboxManager)
-    sandbox_mgr.get_domain_info.return_value = []
-    return SessionEngine(
-        config=config,
-        data_dir=tmp_path,
-        knowledge_dir=tmp_path,
-        git_store=MagicMock(spec=GitStore),
-        session_mgr=session_mgr,
-        skill_catalog=skill_catalog,
-        agent_model=None,
-        sandbox_mgr=sandbox_mgr,
-        credential_registry=CredentialRegistry(),
-        model_factory=lambda _name: TestModel(),
-    )
-
-
-def test_record_tool_call_event_reuses_sentinel_row_for_user_decision(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    sid = engine.session_mgr.create_session().session_id
-
-    initial_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="proxy_domain",
-        args={"domain": "example.com", "command": "curl https://example.com"},
-        detail="[sentinel] reviewing",
-        approval_source="sentinel",
-    )
-    updated_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="proxy_domain",
-        args={"domain": "example.com", "command": "curl https://example.com"},
-        detail="[user: allow]",
-        approval_source="user",
-        approval_verdict="allow",
-    )
-
-    assert updated_tool_id == initial_tool_id
-    events = engine.session_mgr.load_events(sid)
-    assert all("timestamp" in event for event in events)
-    assert _without_timestamps(events) == [
-        {
-            "role": "tool_call",
-            "tool": "proxy_domain",
-            "args": {"domain": "example.com", "command": "curl https://example.com"},
-            "detail": "[user: allow]",
-            "approval_source": "user",
-            "approval_verdict": "allow",
-            "approval_explanation": None,
-            "tool_id": initial_tool_id,
-        }
-    ]
-
-
-def test_record_tool_call_event_reuses_proxy_domain_row_for_queued_reviewing_and_final(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    sid = engine.session_mgr.create_session().session_id
-
-    queued_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="proxy_domain",
-        args={"domain": "example.com"},
-        detail="[sentinel] queued for batched review",
-        approval_source="sentinel",
-        parent_tool_id="tool-1",
-    )
-    reviewing_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="proxy_domain",
-        args={"domain": "example.com"},
-        detail="[sentinel] reviewing",
-        approval_source="sentinel",
-        parent_tool_id="tool-1",
-    )
-    approved_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="proxy_domain",
-        args={"domain": "example.com"},
-        detail="[sentinel: allow] looks fine",
-        approval_source="sentinel",
-        approval_verdict="allow",
-        approval_explanation="looks fine",
-        parent_tool_id="tool-1",
-    )
-
-    assert queued_tool_id == reviewing_tool_id == approved_tool_id
-    events = engine.session_mgr.load_events(sid)
-    assert all("timestamp" in event for event in events)
-    assert _without_timestamps(events) == [
-        {
-            "role": "tool_call",
-            "tool": "proxy_domain",
-            "args": {"domain": "example.com"},
-            "detail": "[sentinel: allow] looks fine",
-            "approval_source": "sentinel",
-            "approval_verdict": "allow",
-            "approval_explanation": "looks fine",
-            "parent_tool_id": "tool-1",
-            "tool_id": queued_tool_id,
-        }
-    ]
-
-
-def test_fork_session_copies_transcript_and_security_context(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    source = engine.session_mgr.create_session(channel_type="matrix", channel_ref="!room:example.com", private=True)
-    sid = source.session_id
-    active = engine.get_or_activate(sid)
-    active.state.title = "Original title"
-    active.state.attributes.private = True
-    active.state.attributes.archived = True
-    active.state.attributes.pinned = True
-    active.state.attributes.favorite = True
-    active.state.approved_operations = ["exec"]
-    active.state.activated_skills = ["web"]
-    active.state.context_grants["web"] = ContextGrant(skill_name="web", domains={"example.com"})
-    active.state.knowledge_last_archive_path = "sessions/archive.json"
-    active.state.knowledge_last_commit_trigger = "manual"
-    active.state.knowledge_last_committed_at = datetime.now(tz=UTC)
-    engine.session_mgr.save_state(active.state)
-    engine.session_mgr.save_usage(
-        sid,
-        usage_mod.UsageTracker(models={"anthropic:test": ModelUsage(input_tokens=11, output_tokens=7)}),
-    )
-    engine.session_mgr.save_sandbox_snapshot(
-        sid,
-        SessionSandboxSnapshot(runtime="kubernetes", status="scaled_down", storage_present=True),
-    )
-    engine.session_mgr.append_events(
-        sid,
-        [
-            {"role": "user", "content": "first"},
-            {"role": "tool_call", "tool": "read", "args": {"path": "README.md"}, "detail": "reading"},
-            {"role": "assistant", "content": "reply one"},
-            {"role": "user", "content": "second"},
-            {"role": "assistant", "content": "reply two"},
-        ],
-    )
-    engine.session_mgr.save_history(
-        sid,
-        [
-            ModelRequest(parts=[UserPromptPart(content="first")]),
-            ModelResponse(parts=[TextPart(content="reply one")]),
-            ModelRequest(parts=[UserPromptPart(content="second")]),
-            ModelResponse(parts=[TextPart(content="reply two")]),
-        ],
-    )
-
-    forked = engine.fork_session(sid, event_index=2, channel_type="web")
-
-    assert forked.session_id != sid
-    assert forked.channel_type == "web"
-    assert forked.channel_ref is None
-    assert forked.title == "Original title (Copy)"
-    assert forked.attributes.private is True
-    assert forked.attributes.archived is False
-    assert forked.attributes.pinned is False
-    assert forked.attributes.favorite is False
-    assert forked.approved_operations == ["exec"]
-    assert forked.activated_skills == ["web"]
-    assert forked.context_grants["web"].domains == {"example.com"}
-    assert forked.knowledge_last_committed_at is None
-    assert forked.knowledge_last_archive_path is None
-    assert forked.knowledge_last_commit_trigger is None
-    assert _without_timestamps(engine.session_mgr.load_events(forked.session_id)) == [
-        {"role": "user", "content": "first"},
-        {"role": "tool_call", "tool": "read", "args": {"path": "README.md"}, "detail": "reading"},
-        {"role": "assistant", "content": "reply one"},
-    ]
-    forked_history = engine.session_mgr.load_history(forked.session_id)
-    assert len(forked_history) == 2
-    assert isinstance(forked_history[0], ModelRequest)
-    assert isinstance(forked_history[1], ModelResponse)
-    assert forked_history[0].parts[0].content == "first"
-    assert forked_history[1].parts[0].content == "reply one"
-    assert engine.session_mgr.load_usage(forked.session_id).models == {}
-    assert engine.session_mgr.load_sandbox_snapshot(forked.session_id) is None
-    assert len(engine.session_mgr.load_events(sid)) == 5
-
-
-def test_handle_token_chunk_promotes_activity_and_broadcasts(tmp_path: Path) -> None:
-    async def _run() -> None:
-        with _patch_sentinel():
-            engine = _make_engine(tmp_path)
-
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-        subscriber = _FakeSubscriber()
-        engine.subscribe(sid, subscriber)
-
-        started_at = datetime.now(tz=UTC)
-        llm_state = LlmRequestState(
-            request_id="req-token",
-            source="agent",
-            model_name="anthropic:claude-haiku-4-5",
-            started_at=started_at,
-            first_text_at=started_at,
-            phase="generating",
-        )
-
-        with patch("carapace.session.engine.note_llm_request_text", return_value=llm_state):
-            await engine._handle_token_chunk(active, "hello")
-
-        assert subscriber.token_chunks == ["hello"]
-        assert active.llm_request_state is not None
-        assert active.llm_request_state.phase == "generating"
-        assert subscriber.llm_activity_updates
-        assert subscriber.llm_activity_updates[-1] is not None
-        assert subscriber.llm_activity_updates[-1].phase == "generating"
-        persisted = engine.session_mgr.load_llm_request_state(sid)
-        assert persisted is not None
-        assert persisted.phase == "generating"
-
-    asyncio.run(_run())
-
-
-def test_handle_thinking_token_chunk_updates_buffer_and_broadcasts(tmp_path: Path) -> None:
-    async def _run() -> None:
-        with _patch_sentinel():
-            engine = _make_engine(tmp_path)
-
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-        subscriber = _FakeSubscriber()
-        engine.subscribe(sid, subscriber)
-
-        started_at = datetime.now(tz=UTC)
-        llm_state = LlmRequestState(
-            request_id="req-thinking",
-            source="agent",
-            model_name="anthropic:claude-haiku-4-5",
-            started_at=started_at,
-            first_thinking_at=started_at,
-            phase="thinking",
-        )
-
-        with patch("carapace.session.engine.note_llm_request_thinking", return_value=llm_state):
-            await engine._handle_thinking_token_chunk(active, "ponder")
-
-        assert subscriber.thinking_chunks == ["ponder"]
-        assert active.llm_request_thinking == {"req-thinking": "ponder"}
-        assert active.llm_request_state is not None
-        assert active.llm_request_state.phase == "thinking"
-        assert subscriber.llm_activity_updates
-        assert subscriber.llm_activity_updates[-1] is not None
-        assert subscriber.llm_activity_updates[-1].phase == "thinking"
-        persisted = engine.session_mgr.load_llm_request_state(sid)
-        assert persisted is not None
-        assert persisted.phase == "thinking"
-
-    asyncio.run(_run())
-
-
-def test_truncate_incomplete_events_keeps_completed_user_approved_exec(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    sid = engine.session_mgr.create_session().session_id
-
-    reviewing_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="exec",
-        args={"command": "ls"},
-        detail="[sentinel] reviewing",
-        approval_source="sentinel",
-    )
-    escalated_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="exec",
-        args={"command": "ls"},
-        detail="[sentinel: escalate] needs approval",
-        approval_source="sentinel",
-        approval_verdict="escalate",
-        approval_explanation="needs approval",
-    )
-    approved_tool_id = engine._record_tool_call_event(
-        sid,
-        tool="exec",
-        args={"command": "ls"},
-        detail="[user approved]",
-        approval_source="user",
-        approval_verdict="allow",
-        approval_explanation="user approved",
-    )
-
-    assert reviewing_tool_id == escalated_tool_id == approved_tool_id
-
-    engine.session_mgr.append_events(
-        sid,
-        [
-            {
-                "role": "tool_result",
-                "tool": "exec",
-                "result": "ok",
-                "exit_code": 0,
-                "tool_id": approved_tool_id,
-            },
-            {
-                "role": "tool_call",
-                "tool": "exec",
-                "args": {"command": "pwd"},
-                "detail": "[safe-list] auto-allowed",
-                "approval_source": "safe-list",
-                "approval_verdict": "allow",
-                "approval_explanation": "auto-allowed",
-                "tool_id": "later-exec",
-            },
-        ],
-    )
-
-    truncated = engine._truncate_incomplete_events(engine.session_mgr.load_events(sid))
-
-    assert all("timestamp" in event for event in truncated)
-    assert _without_timestamps(truncated) == [
-        {
-            "role": "tool_call",
-            "tool": "exec",
-            "args": {"command": "ls"},
-            "detail": "[user approved]",
-            "approval_source": "user",
-            "approval_verdict": "allow",
-            "approval_explanation": "user approved",
-            "tool_id": approved_tool_id,
-        },
-        {
-            "role": "tool_result",
-            "tool": "exec",
-            "result": "ok",
-            "exit_code": 0,
-            "tool_id": approved_tool_id,
-        },
-    ]
-
-
-def test_available_model_entries_override_default_with_metadata(tmp_path: Path):
-    """``available_models`` lists every selectable model; duplicate id in YAML keeps last row metadata."""
-    (tmp_path / "config.yaml").write_text(
-        "agent:\n"
-        "  model: anthropic:alpha\n"
-        "  sentinel_model: anthropic:beta\n"
-        "  title_model: anthropic:gamma\n"
-        "  available_models:\n"
-        "    - provider: anthropic\n"
-        "      name: alpha\n"
-        "      max_input_tokens: 424242\n"
-        "    - provider: anthropic\n"
-        "      name: beta\n"
-        "    - provider: anthropic\n"
-        "      name: gamma\n"
-    )
-    ensure_data_dir(tmp_path)
-    engine = _make_engine(tmp_path)
-    by_id = {e.model_id: e for e in engine.available_model_entries}
-    assert by_id["anthropic:alpha"].max_input_tokens == 424242
-    ids = [e.model_id for e in engine.available_model_entries]
-    assert ids == sorted(ids)
-
-
-def test_user_message_from_self(tmp_path: Path):
-    """Origin subscriber gets from_self=True, others get from_self=False."""
-
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        origin = _FakeSubscriber()
-        other = _FakeSubscriber()
-        engine.subscribe(sid, origin)
-        engine.subscribe(sid, other)
-
-        # Mock run_agent_turn to avoid needing LLM
-        async def _noop_turn(*_a: Any, **_kw: Any) -> str:
-            return "ok"
-
-        with patch("carapace.agent.loop.run_agent_turn", new=_noop_turn):
-            await engine.submit_message(sid, "hello", origin=origin)
-            await asyncio.sleep(0.1)
-
-        assert origin.user_messages == [("hello", True)]
-        assert other.user_messages == [("hello", False)]
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_user_message_no_origin(tmp_path: Path):
-    """When origin is None, all subscribers get from_self=False."""
-
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        sub_a = _FakeSubscriber()
-        sub_b = _FakeSubscriber()
-        engine.subscribe(sid, sub_a)
-        engine.subscribe(sid, sub_b)
-
-        async def _noop_turn(*_a: Any, **_kw: Any) -> str:
-            return "ok"
-
-        with patch("carapace.agent.loop.run_agent_turn", new=_noop_turn):
-            await engine.submit_message(sid, "hi")
-            await asyncio.sleep(0.1)
-
-        assert sub_a.user_messages == [("hi", False)]
-        assert sub_b.user_messages == [("hi", False)]
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-# ---------------------------------------------------------------------------
-# SessionEngine lifecycle tests
-# ---------------------------------------------------------------------------
-
-
-def test_subscribe_duplicate_prevention(tmp_path: Path):
-    """Subscribing the same subscriber twice does not duplicate it."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-        engine.subscribe(sid, sub)
-
-        active = engine.get_active(sid)
-        assert active is not None
-        assert active.subscribers.count(sub) == 1
-
-
-def test_get_active_returns_none_before_activation(tmp_path: Path):
-    """get_active returns None for a session that hasn't been activated."""
-    engine = _make_engine(tmp_path)
-    state = engine.session_mgr.create_session()
-    assert engine.get_active(state.session_id) is None
-
-
-def test_get_or_activate_loads_session(tmp_path: Path):
-    """get_or_activate loads the session from disk and makes it active."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        assert engine.get_active(sid) is None
-        active = engine.get_or_activate(sid)
-        assert active is not None
-        assert active.state.session_id == sid
-        assert engine.get_active(sid) is active
-
-
-def test_get_or_activate_unknown_session_raises(tmp_path: Path):
-    """get_or_activate raises KeyError for a nonexistent session."""
-    engine = _make_engine(tmp_path)
-    with pytest.raises(KeyError):
-        engine.get_or_activate("nonexistent")
-
-
-def test_deactivate_removes_session(tmp_path: Path):
-    """deactivate removes the session from active memory."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        engine.get_or_activate(sid)
-        assert engine.get_active(sid) is not None
-
-        engine.deactivate(sid)
-        assert engine.get_active(sid) is None
-
-
-def test_deactivate_idempotent(tmp_path: Path):
-    """Deactivating an already-deactivated session does not raise."""
-    engine = _make_engine(tmp_path)
-    engine.deactivate("nonexistent")  # should not raise
-
-
-def test_unsubscribe_removes_subscriber(tmp_path: Path):
-    """unsubscribe removes the subscriber from the list."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-        active = engine.get_active(sid)
-        assert sub in active.subscribers
-
-        engine.unsubscribe(sid, sub)
-        assert sub not in active.subscribers
-
-
-def test_unsubscribe_nonexistent_is_safe(tmp_path: Path):
-    """Unsubscribing a subscriber that was never added does not raise."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        engine.get_or_activate(sid)
-        engine.unsubscribe(sid, _FakeSubscriber())  # should not raise
-
-
-def test_unsubscribe_saves_usage_when_last(tmp_path: Path):
-    """Usage is persisted to disk when the last subscriber disconnects."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        # Modify usage so we can detect whether it was saved
-        active = engine.get_active(sid)
-        active.usage_tracker.models["test-model"] = ModelUsage(input_tokens=42)
-
-        engine.unsubscribe(sid, sub)
-
-        # Reload from disk
-        reloaded = engine.session_mgr.load_usage(sid)
-        assert reloaded.total_input == 42
-
-
-def test_submit_message_busy_broadcasts_error(tmp_path: Path):
-    """submit_message rejects with an error if agent is already running."""
-
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        # Simulate a running agent task that never finishes
-        active = engine.get_active(sid)
-        active.agent_task = asyncio.create_task(asyncio.sleep(999))
-
-        await engine.submit_message(sid, "should fail")
-
-        assert any("busy" in e.lower() for e in sub.errors)
-
-        active.agent_task.cancel()
-        with __import__("contextlib").suppress(asyncio.CancelledError):
-            await active.agent_task
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_cancel_stops_task(tmp_path: Path):
-    """submit_cancel cancels the running agent task."""
-
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        engine.subscribe(sid, _FakeSubscriber())
-        active = engine.get_active(sid)
-
-        # Simulate a running agent task
-        active.agent_task = asyncio.create_task(asyncio.sleep(999))
-
-        await engine.submit_cancel(sid)
-        assert active.agent_task is None
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_cancel_persists_interruption_marker(tmp_path: Path):
-    """Cancelled turns are persisted with a terminal assistant message."""
-
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        async def _hanging_turn(*_args: Any, **_kwargs: Any) -> str:
-            await asyncio.sleep(999)
-            return "unreachable"
-
-        with patch("carapace.session.engine.run_agent_turn", new=_hanging_turn):
-            await engine.submit_message(sid, "hello")
-            await asyncio.sleep(0.05)
-            await engine.submit_cancel(sid)
-
-        history = engine.session_mgr.load_history(sid)
-        assert len(history) == 2
-        assert isinstance(history[0], ModelRequest)
-        assert any(isinstance(part, UserPromptPart) and part.content == "hello" for part in history[0].parts)
-        assert isinstance(history[1], ModelResponse)
-        assert any(
-            isinstance(part, TextPart) and part.content == "The previous turn was interrupted before completion."
-            for part in history[1].parts
-        )
-
-        events = engine.session_mgr.load_events(sid)
-        assert all("timestamp" in event for event in events[-2:])
-        assert _without_timestamps(events[-2:]) == [
-            {"role": "user", "content": "hello"},
-            {"role": "assistant", "content": "The previous turn was interrupted before completion."},
-        ]
-        assert sub.cancelled == 1
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_cancel_persists_interrupted_llm_request_log(tmp_path: Path):
-    """Cancelled in-flight LLM requests are saved in the request log as interrupted."""
-
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        engine.subscribe(sid, _FakeSubscriber())
-
-        started_at = datetime(2026, 5, 2, 12, 0, tzinfo=UTC)
-
-        async def _hanging_turn(*_args: Any, **_kwargs: Any) -> str:
-            sink = usage_mod._llm_request_sink.get()
-            assert sink is not None
-            await sink.on_request_started(
-                LlmRequestState(
-                    request_id="req-1",
-                    source="agent",
-                    model_name="anthropic:claude-haiku-4-5",
-                    started_at=started_at,
-                    phase="thinking",
-                    first_thinking_at=started_at,
-                    last_thinking_at=started_at + timedelta(seconds=1),
-                )
-            )
-            await asyncio.sleep(999)
-            return "unreachable"
-
-        with patch("carapace.session.engine.run_agent_turn", new=_hanging_turn):
-            await engine.submit_message(sid, "hello")
-            await asyncio.sleep(0.05)
-            await engine.submit_cancel(sid)
-
-        log = engine.session_mgr.load_llm_request_log(sid)
-        assert len(log.records) == 1
-        record = log.records[0]
-        assert record.request_id == "req-1"
-        assert record.source == "agent"
-        assert record.model_name == "anthropic:claude-haiku-4-5"
-        assert record.outcome == "interrupted"
-        assert record.input_tokens == 0
-        assert record.output_tokens == 0
-        assert record.started_at == started_at
-        assert record.first_thinking_at == started_at
-        assert record.last_thinking_at == started_at + timedelta(seconds=1)
-        assert record.completed_at is None
-        assert engine.session_mgr.load_llm_request_state(sid) is None
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_cancel_noop_when_inactive(tmp_path: Path):
-    """submit_cancel is a no-op when session is not active."""
-
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        await engine.submit_cancel("nonexistent")  # should not raise
-
-    asyncio.run(_run())
-
-
-def test_retry_latest_turn_rewinds_and_restarts(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        engine.session_mgr.save_events(
-            sid,
-            [
-                {"role": "user", "content": "first"},
-                {"role": "assistant", "content": "first answer"},
-                {"role": "user", "content": "second"},
-                {"role": "assistant", "content": "second answer"},
-            ],
-        )
-        engine.session_mgr.save_history(
-            sid,
-            [
-                ModelRequest(parts=[UserPromptPart(content="first")]),
-                ModelResponse(parts=[TextPart(content="first answer")]),
-                ModelRequest(parts=[UserPromptPart(content="second")]),
-                ModelResponse(parts=[TextPart(content="second answer")]),
-            ],
-        )
-
-        async def _fake_run_turn(
-            user_input: str,
-            _deps: Any,
-            message_history: list[Any],
-            **_kwargs: Any,
-        ) -> tuple[list[Any], str, str]:
-            assert user_input == "second"
-            assert len(message_history) == 2
-            assert isinstance(message_history[0], ModelRequest)
-            assert isinstance(message_history[1], ModelResponse)
-            return (
-                [
-                    *message_history,
-                    ModelRequest(parts=[UserPromptPart(content=user_input)]),
-                    ModelResponse(parts=[TextPart(content="retried answer")]),
-                ],
-                "retried answer",
-                "",
-            )
-
-        with patch("carapace.session.engine.run_agent_turn", new=_fake_run_turn):
-            await engine.retry_latest_turn(sid, origin=sub)
-            active = engine.get_active(sid)
-            assert active is not None and active.agent_task is not None
-            await active.agent_task
-
-        events = engine.session_mgr.load_events(sid)
-        assert _without_timestamps(events) == [
-            {"role": "user", "content": "first"},
-            {"role": "assistant", "content": "first answer"},
-            {"role": "user", "content": "second"},
-            {"role": "assistant", "content": "retried answer"},
-        ]
-
-        history = engine.session_mgr.load_history(sid)
-        assert len(history) == 4
-        assert isinstance(history[-1], ModelResponse)
-        assert any(isinstance(part, TextPart) and part.content == "retried answer" for part in history[-1].parts)
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_retry_latest_turn_after_failure_uses_terminal_marker_history(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        sid = engine.session_mgr.create_session().session_id
-
-        engine.session_mgr.save_events(
-            sid,
-            [
-                {"role": "user", "content": "hello"},
-                {"role": "assistant", "content": "The previous turn failed before completion."},
-            ],
-        )
-        engine.session_mgr.save_history(
-            sid,
-            [
-                ModelRequest(parts=[UserPromptPart(content="hello")]),
-                ModelResponse(parts=[TextPart(content="The previous turn failed before completion.")]),
-            ],
-        )
-
-        async def _fake_run_turn(
-            user_input: str,
-            _deps: Any,
-            message_history: list[Any],
-            **_kwargs: Any,
-        ) -> tuple[list[Any], str, str]:
-            assert user_input == "hello"
-            assert message_history == []
-            return (
-                [
-                    ModelRequest(parts=[UserPromptPart(content=user_input)]),
-                    ModelResponse(parts=[TextPart(content="recovered")]),
-                ],
-                "recovered",
-                "",
-            )
-
-        with patch("carapace.session.engine.run_agent_turn", new=_fake_run_turn):
-            await engine.retry_latest_turn(sid)
-            active = engine.get_active(sid)
-            assert active is not None and active.agent_task is not None
-            await active.agent_task
-
-        events = engine.session_mgr.load_events(sid)
-        assert _without_timestamps(events) == [
-            {"role": "user", "content": "hello"},
-            {"role": "assistant", "content": "recovered"},
-        ]
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_reset_to_turn_rewinds_later_turns(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        sid = engine.session_mgr.create_session().session_id
-
-        engine.session_mgr.save_events(
-            sid,
-            [
-                {"role": "user", "content": "first"},
-                {"role": "assistant", "content": "first answer"},
-                {"role": "user", "content": "second"},
-                {"role": "assistant", "content": "second answer"},
-            ],
-        )
-        engine.session_mgr.save_history(
-            sid,
-            [
-                ModelRequest(parts=[UserPromptPart(content="first")]),
-                ModelResponse(parts=[TextPart(content="first answer")]),
-                ModelRequest(parts=[UserPromptPart(content="second")]),
-                ModelResponse(parts=[TextPart(content="second answer")]),
-            ],
-        )
-
-        reset_applied = await engine.reset_to_turn(sid, 1)
-
-        assert reset_applied is True
-
-        events = engine.session_mgr.load_events(sid)
-        assert _without_timestamps(events) == [
-            {"role": "user", "content": "first"},
-            {"role": "assistant", "content": "first answer"},
-        ]
-
-        history = engine.session_mgr.load_history(sid)
-        assert len(history) == 2
-        assert isinstance(history[0], ModelRequest)
-        assert isinstance(history[1], ModelResponse)
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_reset_to_turn_rejects_unknown_target(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        sid = engine.session_mgr.create_session().session_id
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        engine.session_mgr.save_events(
-            sid,
-            [
-                {"role": "user", "content": "hello"},
-                {"role": "assistant", "content": "world"},
-            ],
-        )
-
-        reset_applied = await engine.reset_to_turn(sid, 99)
-
-        assert reset_applied is False
-        assert sub.errors == ["Unknown reset target"]
-        assert sub.error_events == [("Unknown reset target", False)]
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_history_for_completed_turn_count_excludes_trailing_incomplete_request(tmp_path: Path):
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-
-    history = [
-        ModelRequest(parts=[UserPromptPart(content="first")]),
-        ModelResponse(parts=[TextPart(content="first answer")]),
-        ModelRequest(parts=[UserPromptPart(content="second")]),
-    ]
-
-    assert engine._completed_model_turn_end_indexes(history) == [1]
-    assert engine._history_for_completed_turn_count(history, 2) == history[:2]
-
-
-def test_handle_slash_command_session(tmp_path: Path):
-    """handle_slash_command /session returns session metadata."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/session")
-            assert result is not None
-            assert result["command"] == "session"
-            assert result["data"]["session_id"] == sid
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_unknown(tmp_path: Path):
-    """handle_slash_command returns None for unknown commands."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            assert await engine.handle_slash_command(sid, "/nonexistent") is None
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_retitle_sets_title(tmp_path: Path):
-    """``/retitle TEXT`` stores the title and returns a message."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/retitle Hello world")
-            assert result is not None
-            assert result["command"] == "retitle"
-            assert "Hello world" in result["data"]["message"]
-            assert active.state.title == "Hello world"
-            reloaded = engine.session_mgr.load_state(sid)
-            assert reloaded is not None
-            assert reloaded.title == "Hello world"
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_retitle_regenerates(tmp_path: Path):
-    """``/retitle`` with no args runs title generation."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-        engine.session_mgr.append_events(sid, [{"role": "user", "content": "talk about cats"}])
-
-        async def _run() -> None:
-            with patch(
-                "carapace.session.engine.generate_title",
-                new=AsyncMock(return_value="📌 Cats chat"),
-            ):
-                result = await engine.handle_slash_command(sid, "/retitle")
-            assert result is not None
-            assert result["command"] == "retitle"
-            assert "Cats chat" in result["data"]["message"]
-            assert active.state.title == "📌 Cats chat"
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_budget_sets_and_clears(tmp_path: Path):
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            initial = await engine.handle_slash_command(sid, "/budget")
-            assert initial is not None
-            assert "usage_hint" in initial["data"]
-            assert "/budget input N" in initial["data"]["usage_hint"]
-            assert "/budget tools N" in initial["data"]["usage_hint"]
-
-            set_result = await engine.handle_slash_command(sid, "/budget input 1000")
-            assert set_result is not None
-            assert set_result["command"] == "budget"
-            assert set_result["data"]["message"] == "Set input token budget to 1,000 tokens."
-            assert set_result["data"]["gauges"][0]["key"] == "input"
-
-            cleared = await engine.handle_slash_command(sid, "/budget input 0")
-            assert cleared is not None
-            assert cleared["data"]["message"] == "Cleared input token budget."
-            assert cleared["data"]["gauges"] == []
-
-            reloaded = engine.session_mgr.load_state(sid)
-            assert reloaded is not None
-            assert reloaded.budget.input_tokens is None
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_budget_sets_tool_call_limit(tmp_path: Path):
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/budget tools 3")
-
-            assert result is not None
-            assert result["command"] == "budget"
-            assert result["data"]["message"] == "Set tool call budget to 3 tool calls."
-            assert result["data"]["gauges"][0]["key"] == "tool_calls"
-
-            reloaded = engine.session_mgr.load_state(sid)
-            assert reloaded is not None
-            assert reloaded.budget.tool_calls == 3
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_budget_accepts_k_and_m_suffixes(tmp_path: Path):
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            input_result = await engine.handle_slash_command(sid, "/budget input 100k")
-            assert input_result is not None
-            assert input_result["data"]["message"] == "Set input token budget to 100,000 tokens."
-
-            output_result = await engine.handle_slash_command(sid, "/budget output 2M")
-            assert output_result is not None
-            assert output_result["data"]["message"] == "Set output token budget to 2,000,000 tokens."
-
-            reloaded = engine.session_mgr.load_state(sid)
-            assert reloaded is not None
-            assert reloaded.budget.input_tokens == 100_000
-            assert reloaded.budget.output_tokens == 2_000_000
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_help_lists_budget(tmp_path: Path):
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/help")
-            assert result is not None
-            commands = result["data"]["commands"]
-            assert any(item["command"] == "/budget" for item in commands)
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_usage_includes_tool_call_total(tmp_path: Path):
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-        active.usage_tracker.record_tool_call()
-        active.usage_tracker.record_tool_call()
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/usage")
-
-            assert result is not None
-            assert result["command"] == "usage"
-            assert result["data"]["total_tool_calls"] == 2
-
-        asyncio.run(_run())
-
-
-def test_turn_usage_payload_contains_budget_gauges_without_agent_usage(tmp_path: Path):
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session(budget=SessionBudget(input_tokens=1_000))
-        active = engine.get_or_activate(state.session_id)
-
-        payload = engine._turn_usage_payload(active)
-
-        assert payload is not None
-        assert payload.budget_gauges[0].key == "input"
-        assert payload.budget_gauges[0].current_value == "0 tokens"
-
-
-def test_turn_usage_payload_includes_reasoning_metrics(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        active = engine.get_or_activate(state.session_id)
-        started_at = datetime.now(tz=UTC)
-        active.llm_request_log.records.append(
-            LlmRequestRecord(
-                ts=started_at + timedelta(seconds=5),
-                source="agent",
-                model_name="anthropic:claude-haiku-4-5",
-                input_tokens=100,
-                output_tokens=20,
-                started_at=started_at,
-                first_thinking_at=started_at + timedelta(seconds=1),
-                last_thinking_at=started_at + timedelta(seconds=2),
-                first_text_at=started_at + timedelta(seconds=3),
-                completed_at=started_at + timedelta(seconds=5),
-                reasoning_tokens=42,
-            )
-        )
-
-        payload = engine._turn_usage_payload(active)
-
-        assert payload is not None
-        assert payload.ttft_ms == 3000
-        assert payload.reasoning_duration_ms == 2000
-        assert payload.total_duration_ms == 5000
-        assert payload.reasoning_tokens == 42
-
-
-def test_activate_clears_stale_llm_request_state(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        engine.session_mgr.save_llm_request_state(
-            state.session_id,
-            LlmRequestState(
-                request_id="stale",
-                source="agent",
-                model_name="anthropic:claude-haiku-4-5",
-                started_at=datetime.now(tz=UTC),
-                phase="thinking",
-            ),
-        )
-
-        active = engine.get_or_activate(state.session_id)
-
-        assert active.llm_request_state is None
-        assert engine.session_mgr.load_llm_request_state(state.session_id) is None
-
-
-def test_llm_request_recording_persists_request_level_thinking_event(tmp_path: Path) -> None:
-    async def _run() -> None:
-        with _patch_sentinel():
-            engine = _make_engine(tmp_path)
-            state = engine.session_mgr.create_session()
-            active = engine.get_or_activate(state.session_id)
-            started_at = datetime.now(tz=UTC)
-            request_state = LlmRequestState(
-                request_id="req-1",
-                source="agent",
-                model_name="anthropic:claude-haiku-4-5",
-                started_at=started_at,
-                phase="thinking",
-                first_thinking_at=started_at,
-            )
-            record = LlmRequestRecord(
-                ts=started_at + timedelta(seconds=2),
-                request_id="req-1",
-                source="agent",
-                model_name="anthropic:claude-haiku-4-5",
-                started_at=started_at,
-                first_thinking_at=started_at,
-                first_text_at=started_at + timedelta(seconds=1),
-                completed_at=started_at + timedelta(seconds=2),
-                reasoning_tokens=17,
-            )
-
-            with engine.llm_request_recording(active):
-                observer = usage_mod._llm_request_sink.get()
-                assert observer is not None
-                await observer.on_request_started(request_state)
-                active.llm_request_thinking["req-1"] = "first thought"
-                await observer.on_request_completed(record)
-
-            events = engine.session_mgr.load_events(state.session_id)
-            assert "timestamp" in events[-1]
-            assert _without_timestamp(events[-1]) == {
-                "role": "thinking",
-                "content": "first thought",
-                "request_id": "req-1",
-                "reasoning_duration_ms": 1000,
-                "reasoning_tokens": 17,
-            }
-
-    asyncio.run(_run())
-
-
-def test_llm_request_recording_persists_timing_for_tool_only_thinking_event(tmp_path: Path) -> None:
-    async def _run() -> None:
-        with _patch_sentinel():
-            engine = _make_engine(tmp_path)
-            state = engine.session_mgr.create_session()
-            active = engine.get_or_activate(state.session_id)
-            started_at = datetime.now(tz=UTC)
-            request_state = LlmRequestState(
-                request_id="req-tool",
-                source="agent",
-                model_name="anthropic:claude-haiku-4-5",
-                started_at=started_at,
-                phase="thinking",
-                first_thinking_at=started_at,
-            )
-            record = LlmRequestRecord(
-                ts=started_at + timedelta(seconds=3),
-                request_id="req-tool",
-                source="agent",
-                model_name="anthropic:claude-haiku-4-5",
-                started_at=started_at,
-                first_thinking_at=started_at,
-                completed_at=started_at + timedelta(seconds=3),
-                reasoning_tokens=9,
-            )
-
-            with engine.llm_request_recording(active):
-                observer = usage_mod._llm_request_sink.get()
-                assert observer is not None
-                await observer.on_request_started(request_state)
-                active.llm_request_thinking["req-tool"] = "tool-only thought"
-                await observer.on_request_completed(record)
-
-            events = engine.session_mgr.load_events(state.session_id)
-            assert "timestamp" in events[-1]
-            assert _without_timestamp(events[-1]) == {
-                "role": "thinking",
-                "content": "tool-only thought",
-                "request_id": "req-tool",
-                "reasoning_duration_ms": 3000,
-                "reasoning_tokens": 9,
-            }
-
-    asyncio.run(_run())
-
-
-def test_submit_message_budget_exhausted_broadcasts_error(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session(budget=SessionBudget(input_tokens=100))
-        sid = state.session_id
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-        active = engine.get_active(sid)
-        assert active is not None
-        active.usage_tracker.models["test-model"] = ModelUsage(input_tokens=120)
-
-        with patch("carapace.session.engine.run_agent_turn", new=AsyncMock()) as mocked_turn:
-            await engine.submit_message(sid, "hello")
-            await asyncio.sleep(0.1)
-
-        mocked_turn.assert_not_awaited()
-        assert any("Session budget reached" in err for err in sub.errors)
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_message_tool_call_budget_exhausted_broadcasts_error(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session(budget=SessionBudget(tool_calls=1))
-        sid = state.session_id
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        async def _fail_on_second_tool_call(
-            _user_input: str,
-            deps,
-            _message_history: list[Any],
-            *_args: Any,
-            **_kwargs: Any,
-        ):
-            callback = deps.tool_call_callback
-            assert callback is not None
-            callback("read", {"path": "README.md"}, "[safe-list] auto-allowed", "safe-list", "allow", "ok")
-            callback("read", {"path": "README.md"}, "[safe-list] auto-allowed", "safe-list", "allow", "ok")
-            return [], "done", ""
-
-        with patch("carapace.session.engine.run_agent_turn", new=_fail_on_second_tool_call):
-            await engine.submit_message(sid, "hello")
-            await asyncio.sleep(0.1)
-
-        active = engine.get_active(sid)
-        assert active is not None
-        assert active.usage_tracker.tool_calls == 1
-        assert any("Session budget reached: tool calls 1 tool calls / 1 tool calls" in err for err in sub.errors)
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_message_refreshes_sandbox_once_after_completed_turn(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        async def _complete_turn(*_args: Any, **_kwargs: Any):
-            return [], "done", "thinking"
-
-        with patch("carapace.session.engine.run_agent_turn", new=_complete_turn):
-            await engine.submit_message(sid, "hello")
-            await asyncio.sleep(0.1)
-
-        engine._sandbox_mgr.refresh_sandbox_snapshot.assert_awaited_once_with(sid, measure_usage=True)
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_message_refresh_failure_does_not_block_completed_turn(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        async def _complete_turn(*_args: Any, **_kwargs: Any):
-            return [], "done", "thinking"
-
-        engine._sandbox_mgr.refresh_sandbox_snapshot.side_effect = RuntimeError("snapshot refresh failed")
-
-        with patch("carapace.session.engine.run_agent_turn", new=_complete_turn):
-            await engine.submit_message(sid, "hello")
-            await asyncio.sleep(0.1)
-
-        assert sub.errors == []
-        event = engine.session_mgr.load_events(sid)[-1]
-        assert "timestamp" in event
-        assert _without_timestamp(event) == {"role": "assistant", "content": "done"}
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_message_budget_exceeded_persists_history(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session(budget=SessionBudget(input_tokens=1_000))
-        sid = state.session_id
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        async def _fail_run_agent_turn(*_args: Any, on_messages_snapshot=None, **_kwargs: Any):
-            snapshot = [ModelRequest(parts=[UserPromptPart(content="hello")])]
-            if on_messages_snapshot is not None:
-                on_messages_snapshot(snapshot)
-            raise SessionBudgetExceededError("Session budget reached: input 1.0k tokens / 1.0k tokens", gauges=[])
-
-        with patch("carapace.session.engine.run_agent_turn", new=_fail_run_agent_turn):
-            await engine.submit_message(sid, "hello")
-            await asyncio.sleep(0.1)
-
-        history = engine.session_mgr.load_history(sid)
-        assert history
-        assert isinstance(history[0], ModelRequest)
-        assert any(isinstance(part, UserPromptPart) and part.content == "hello" for part in history[0].parts)
-        assert isinstance(history[-1], ModelResponse)
-        assert any(
-            isinstance(part, TextPart) and part.content == "Session budget reached: input 1.0k tokens / 1.0k tokens"
-            for part in history[-1].parts
-        )
-
-        events = engine.session_mgr.load_events(sid)
-        assert "timestamp" in events[-1]
-        assert _without_timestamp(events[-1]) == {
-            "role": "assistant",
-            "content": "Session budget reached: input 1.0k tokens / 1.0k tokens",
-        }
-        assert any("Session budget reached" in err for err in sub.errors)
-        assert any(
-            detail.startswith("Session budget reached") and turn_terminal for detail, turn_terminal in sub.error_events
-        )
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_submit_message_unexpected_output_marks_terminal_error(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        sid = engine.session_mgr.create_session().session_id
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        unexpected = "Unexpected agent output type: {'message': 'bad'}"
-
-        async def _unexpected_run_agent_turn(*_args: Any, **_kwargs: Any):
-            return (
-                [
-                    ModelRequest(parts=[UserPromptPart(content="hello")]),
-                    ModelResponse(parts=[TextPart(content="placeholder")]),
-                ],
-                unexpected,
-                "",
-            )
-
-        with patch("carapace.session.engine.run_agent_turn", new=_unexpected_run_agent_turn):
-            await engine.submit_message(sid, "hello")
-            active = engine.get_active(sid)
-            assert active is not None and active.agent_task is not None
-            await active.agent_task
-
-        events = engine.session_mgr.load_events(sid)
-        assert events[-1]["role"] == "assistant"
-        assert events[-1]["content"] == unexpected
-        assert sub.done_messages == []
-        assert sub.error_events == [(unexpected, True)]
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_evaluate_with_usage_limit_exceeded_escalates_to_user(tmp_path: Path):
-    async def _run() -> None:
-        session = SessionSecurity("test-session", audit_dir=tmp_path)
-        sentinel = MagicMock(spec=Sentinel)
-        sentinel.evaluate_tool_call = AsyncMock(
-            side_effect=UsageLimitExceeded("The next request would exceed the request_limit of 5")
-        )
-
-        with pytest.raises(ApprovalRequired) as exc_info:
-            await security_mod.evaluate_with(
-                session,
-                sentinel,
-                "use_skill",
-                {"skill_name": "paperless"},
-            )
-
-        metadata = exc_info.value.metadata
-        assert metadata["tool"] == "use_skill"
-        assert metadata["args"] == {"skill_name": "paperless"}
-        assert metadata["risk_level"] == "high"
-        assert isinstance(metadata["sentinel_verdict"], SentinelVerdict)
-        assert metadata["sentinel_verdict"].decision == "escalate"
-        assert "request limit" in metadata["explanation"].lower()
-
-        entry = session.action_log[-1]
-        assert isinstance(entry, ToolCallEntry)
-        assert entry.decision == "escalated"
-        assert entry.tool == "use_skill"
-        assert entry.explanation is not None
-        assert "request limit" in entry.explanation.lower()
-
-    asyncio.run(_run())
-
-
-def test_generate_title_skips_when_budget_exhausted(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session(budget=SessionBudget(input_tokens=100))
-        active = engine.get_or_activate(state.session_id)
-        active.usage_tracker.models["test-model"] = ModelUsage(input_tokens=120)
-
-        with patch("carapace.session.engine.generate_title", new=AsyncMock(return_value="ignored")) as mocked:
-            title = await engine._generate_title(active, [{"role": "user", "content": "hello"}])
-
-        assert title == ""
-        mocked.assert_not_awaited()
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_generate_title_persists_usage_and_broadcasts_usage(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session(budget=SessionBudget(cost_usd=Decimal("5.00")))
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-        sub = _FakeSubscriber()
-        engine.subscribe(sid, sub)
-
-        async def _fake_generate_title(*_args: Any, usage_tracker, **_kwargs: Any) -> str:
-            usage_tracker.record(
-                "anthropic:claude-haiku-4-5",
-                "title",
-                RunUsage(input_tokens=10, output_tokens=5, requests=1),
-            )
-            return "📌 hello"
-
-        with patch("carapace.session.engine.generate_title", new=AsyncMock(side_effect=_fake_generate_title)):
-            title = await engine._generate_title(active, [{"role": "user", "content": "hello"}])
-
-        assert title == "📌 hello"
-        stored_usage = engine.session_mgr.load_usage(sid)
-        assert stored_usage.categories["title"].input_tokens == 10
-        assert sub.title_updates
-        assert sub.title_updates[0][0] == "📌 hello"
-        assert sub.title_updates[0][1] is not None
-        assert sub.title_updates[0][1].budget_gauges[0].key == "cost"
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_generate_title_records_titler_request_log(tmp_path: Path):
-    async def _run() -> None:
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-        started_at = datetime.now(tz=UTC)
-
-        async def _fake_generate_title(*_args: Any, **_kwargs: Any) -> str:
-            observer = usage_mod._llm_request_sink.get()
-            assert observer is not None
-
-            request_state = LlmRequestState(
-                request_id="req-title",
-                source="titler",
-                model_name="anthropic:claude-haiku-4-5",
-                started_at=started_at,
-            )
-            record = LlmRequestRecord(
-                ts=started_at + timedelta(seconds=1),
-                request_id="req-title",
-                source="titler",
-                model_name="anthropic:claude-haiku-4-5",
-                started_at=started_at,
-                completed_at=started_at + timedelta(seconds=1),
-            )
-
-            await observer.on_request_started(request_state)
-            await observer.on_request_completed(record)
-            return "📌 hello"
-
-        with patch("carapace.session.engine.generate_title", new=AsyncMock(side_effect=_fake_generate_title)):
-            title = await engine._generate_title(active, [{"role": "user", "content": "hello"}])
-
-        assert title == "📌 hello"
-        assert active.llm_request_log.records[-1].request_id == "req-title"
-        assert active.llm_request_log.records[-1].source == "titler"
-
-        stored_log = engine.session_mgr.load_llm_request_log(sid)
-        assert stored_log.records[-1].request_id == "req-title"
-        assert stored_log.records[-1].source == "titler"
-
-    with _patch_sentinel():
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_inactive_session(tmp_path: Path):
-    """handle_slash_command returns None for a session that isn't active."""
-    engine = _make_engine(tmp_path)
-    state = engine.session_mgr.create_session()
-
-    async def _run() -> None:
-        assert await engine.handle_slash_command(state.session_id, "/session") is None
-
-    asyncio.run(_run())
-
-
-def test_handle_slash_command_reload(tmp_path: Path):
-    """handle_slash_command /reload calls reset_session and returns success."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/reload")
-            assert result is not None
-            assert result["command"] == "reload"
-            assert "reset" in result["data"]["message"].lower() or "fresh" in result["data"]["message"].lower()
-            engine._sandbox_mgr.reset_session.assert_called_once_with(sid)
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_model_sets_all_three(tmp_path: Path):
-    """``/model NAME`` applies the same id to agent, sentinel, and title."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/model openai:gpt-4o")
-            assert result is not None
-            assert result["command"] == "model"
-            assert "models" in result["data"]
-            for key in ("agent", "sentinel", "title"):
-                assert result["data"]["models"][key]["current"] == "openai:gpt-4o"
-            assert active.agent_model_name == "openai:gpt-4o"
-            assert active.sentinel_model_name == "openai:gpt-4o"
-            assert active.title_model_name == "openai:gpt-4o"
-
-        asyncio.run(_run())
-
-
-def test_handle_slash_command_model_agent_only(tmp_path: Path):
-    """``/model-agent`` changes only the agent model."""
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        active = engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/model-agent openai:gpt-4o")
-            assert result is not None
-            assert result["command"] == "model-agent"
-            assert result["data"]["current"] == "openai:gpt-4o"
-            assert active.agent_model_name == "openai:gpt-4o"
-            assert active.sentinel_model_name is None
-            assert active.title_model_name is None
-
-        asyncio.run(_run())
-
-
-def test_model_overrides_persist_across_restart(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-        sid = state.session_id
-        engine.get_or_activate(sid)
-
-        async def _run() -> None:
-            result = await engine.handle_slash_command(sid, "/model openai:gpt-4o")
-            assert result is not None
-            assert result["command"] == "model"
-
-        asyncio.run(_run())
-
-    persisted = engine.session_mgr.load_state(sid)
-    assert persisted is not None
-    assert persisted.agent_model_name == "openai:gpt-4o"
-    assert persisted.sentinel_model_name == "openai:gpt-4o"
-    assert persisted.title_model_name == "openai:gpt-4o"
-
-    with _patch_sentinel():
-        restarted = _make_engine(tmp_path)
-        active = restarted.get_or_activate(sid)
-        deps = restarted._build_deps(active)
-
-    assert active.agent_model_name == "openai:gpt-4o"
-    assert active.sentinel_model_name == "openai:gpt-4o"
-    assert active.title_model_name == "openai:gpt-4o"
-    assert isinstance(deps.agent_model, TestModel)
-    assert deps.agent_model_id == "openai:gpt-4o"
-    assert active.agent_model is deps.agent_model
-    assert active.sentinel is not None
-    active.sentinel.set_model.assert_called_once_with("openai:gpt-4o")
-
-
-def test_invalid_model_overrides_fall_back_on_restart(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        state = engine.session_mgr.create_session()
-
-    state.agent_model_name = "openai:missing-agent"
-    state.sentinel_model_name = "openai:missing-sentinel"
-    state.title_model_name = "openai:missing-title"
-    engine.session_mgr.save_state(state)
-
-    with _patch_sentinel() as sentinel_cls, patch("carapace.session.engine.logger.warning") as warning_mock:
-        sentinel_instance = sentinel_cls.return_value
-
-        def _set_model(name: str) -> None:
-            if name == "openai:missing-sentinel":
-                raise ValueError("missing sentinel model")
-
-        sentinel_instance.set_model.side_effect = _set_model
-
-        restarted = _make_engine(tmp_path)
-        restarted._resolve_model = MagicMock(
-            side_effect=lambda name: (
-                TestModel()
-                if name == restarted._config.agent.model
-                else (_ for _ in ()).throw(ValueError("missing agent model"))
-            )
-        )
-
-        active = restarted.get_or_activate(state.session_id)
-        deps = restarted._build_deps(active)
-
-    assert active.agent_model_name is None
-    assert active.sentinel_model_name is None
-    assert active.title_model_name is None
-    assert isinstance(deps.agent_model, TestModel)
-    assert deps.agent_model_id == restarted._config.agent.model
-
-    persisted = restarted.session_mgr.load_state(state.session_id)
-    assert persisted is not None
-    assert persisted.agent_model_name is None
-    assert persisted.sentinel_model_name is None
-    assert persisted.title_model_name is None
-
-    sentinel_instance.set_model.assert_any_call("openai:missing-sentinel")
-    sentinel_instance.set_model.assert_any_call(restarted._config.agent.sentinel_model)
-    assert warning_mock.call_count == 3
-
-
-def test_non_slash_user_message_count_ignores_slash_lines() -> None:
-    events: list[dict[str, Any]] = [
-        {"role": "user", "content": "/model-agent openai:gpt-4o"},
-        {"role": "command", "command": "model-agent", "data": {}},
-        {"role": "user", "content": "hello"},
-    ]
-    assert _non_slash_user_message_count(events) == 1
-
-
-def test_non_slash_user_message_count_plain_users() -> None:
-    events: list[dict[str, Any]] = [
-        {"role": "user", "content": "a"},
-        {"role": "user", "content": "b"},
-        {"role": "user", "content": "c"},
-    ]
-    assert _non_slash_user_message_count(events) == 3
-
-
-def test_truncate_incomplete_model_history_drops_dangling_tool_tail(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        messages = [
-            ModelRequest(parts=[UserPromptPart(content="hello")]),
-            ModelResponse(parts=[ToolCallPart(tool_name="cmd", args={}, tool_call_id="call-1")]),
-            ModelRequest(parts=[ToolReturnPart(tool_name="cmd", content="ok", tool_call_id="call-1")]),
-            ModelResponse(parts=[TextPart(content="done")]),
-            ModelResponse(parts=[ToolCallPart(tool_name="cmd", args={}, tool_call_id="call-2")]),
-        ]
-
-        truncated = engine._truncate_incomplete_model_history(messages)
-
-        assert len(truncated) == 4
-        assert isinstance(truncated[-1], ModelResponse)
-        assert truncated[-1].parts[0].part_kind == "text"
-
-
-def test_truncate_incomplete_model_history_keeps_complete_pairs(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        messages = [
-            ModelRequest(parts=[UserPromptPart(content="hello")]),
-            ModelResponse(parts=[ToolCallPart(tool_name="cmd", args={}, tool_call_id="call-1")]),
-            ModelRequest(parts=[ToolReturnPart(tool_name="cmd", content="ok", tool_call_id="call-1")]),
-            ModelResponse(parts=[TextPart(content="done")]),
-        ]
-
-        truncated = engine._truncate_incomplete_model_history(messages)
-        assert truncated == messages
-
-
-def test_truncate_incomplete_events_drops_dangling_tool_tail(tmp_path: Path) -> None:
-    with _patch_sentinel():
-        engine = _make_engine(tmp_path)
-        events: list[dict[str, Any]] = [
-            {"role": "user", "content": "hello"},
-            {"role": "tool_call", "tool": "shell", "args": {"command": "echo ok"}, "detail": "run"},
-            {"role": "tool_result", "tool": "shell", "result": "ok", "exit_code": 0},
-            {"role": "assistant", "content": "done"},
-            {"role": "tool_call", "tool": "shell", "args": {"command": "sleep 10"}, "detail": "run"},
-        ]
-
-        truncated = engine._truncate_incomplete_events(events)
-
-        assert len(truncated) == 4
-        assert truncated[-1]["role"] == "assistant"
